requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
aiohttp==3.9.1
httpx==0.25.2
jsonschema==4.20.0
celery[redis]==5.3.4
redis==5.0.1
websockets==12.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1